FM2-400: Separated model and operations of Lastn
